some fixes with filters

+++ src/Components/Main/TaskTable.js
@@ -1,39 +1,44 @@
 import React, { Component } from 'react';
 // import EditTask from './EditTask'
+// import $ from "jquery";
 const firebase = require('firebase');
 
 class TaskTable extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            tasks: []
-        };
-    }
+    // constructor(props) {
+    //     super(props);
+    //     this.state = {
+    //         tasks: []
+    //     };
+    // }
 
-    componentDidMount() {
-        firebase.firestore().collection('tasks').onSnapshot(dataUpdate => {
-            const tasks = dataUpdate.docs.map(doc => {
-                const data = doc.data();
-                data['id'] = doc.id;
-                return data;
-            });
-            this.setState({ tasks })
-        });
-    }
+    // componentDidMount() {
+    //     firebase.firestore().collection('tasks').onSnapshot(dataUpdate => {
+    //         const tasks = dataUpdate.docs.map(doc => {
+    //             const data = doc.data();
+    //             data['id'] = doc.id;
+    //             return data;
+    //         });
+    //         this.setState({ tasks })
+    //     });
+    // }
 
     deleteTask = (taskId) => {
-
-        try {
-            firebase.firestore().collection('tasks').doc(taskId).delete().then(() => {
-                console.log("Task successfully deleted!");
-            });
-        } catch {
-            alert("some error happened upon deleting task :/ ");
+        let acceptance = window.confirm("Press OK if you want to proceed removing chosen task");
+        if (acceptance) {
+            try {
+                firebase.firestore().collection('tasks').doc(taskId).delete().then(() => {
+                    console.log("Task successfully deleted!");
+                });
+            } catch {
+                alert("some error happened upon deleting task :/ ");
+            }
         }
+        // document.getElementById('root').createElement(<div class="modal-backdrop fade show"></div>)
+
     }
 
     renderTableBody = () => {
-        return this.state.tasks.map((task) => {
+        return this.props.tasks.map((task) => {
             const { id, name, status, description } = task;
             return (
                 <tr key={id}>
@@ -41,7 +46,7 @@ class TaskTable extends Component {
                     <td>{name}</td>
                     <td>{description}</td>
                     <td className="text-nowrap">
-                        <i className="btn btn-sm btn-outline-light fas fa-edit mr-2"></i>
+                        <i className="btn btn-sm btn-outline-light fas fa-edit mr-2" title="Edit"></i>
                         <i onClick={() => { this.deleteTask(id) }} title="Delete" className="btn btn-sm btn-outline-light fas fa-trash-alt"></i>
                     </td>
                 </tr>
